Moved interface to container and added mocks
