+++ pkg/gofr/container/container.go
@@ -19,8 +19,6 @@ import (
 	"strings"
 	"time"
 
-	"gofr.dev/pkg/gofr/datasource/opentsdb"
-
 	_ "github.com/go-sql-driver/mysql" // This is required to be blank import
 
 	"gofr.dev/pkg/gofr/config"
@@ -60,7 +58,7 @@ type Container struct {
 	Mongo      Mongo
 	Solr       Solr
 	DGraph     Dgraph
-	Opentsdb   opentsdb.OpentsDBClient
+	Opentsdb   OpentsDBClient
 
 	KVStore KVStore
