+++ pkg/gofr/container/health.go
@@ -65,6 +65,7 @@ func checkExternalDBHealth(ctx context.Context, c *Container, healthMap map[stri
 		"clickHouse": c.Clickhouse,
 		"kv-store":   c.KVStore,
 		"dgraph":     c.DGraph,
+		"opentsdb":   c.Opentsdb,
 	}
 
 	for name, service := range services {
