+++ pkg/gofr/container/mock_container.go
@@ -6,8 +6,6 @@ import (
 	"net/http"
 	"testing"
 
-	"gofr.dev/pkg/gofr/datasource/opentsdb"
-
 	"go.uber.org/mock/gomock"
 
 	"gofr.dev/pkg/gofr/datasource"
@@ -26,7 +24,7 @@ type Mocks struct {
 	Mongo       *MockMongo
 	KVStore     *MockKVStore
 	DGraph      *MockDgraph
-	Opentsdb    *opentsdb.MockOpentsdbProvider
+	Opentsdb    *MockOpentsdbProvider
 	File        *file.MockFileSystemProvider
 	HTTPService *service.MockHTTP
 	Metrics     *MockMetrics
@@ -86,7 +84,7 @@ func NewMockContainer(t *testing.T, options ...options) (*Container, *Mocks) {
 	dgraphMock := NewMockDgraph(ctrl)
 	container.DGraph = dgraphMock
 
-	opentsdbMock := opentsdb.NewMockOpentsdbProvider(ctrl)
+	opentsdbMock := NewMockOpentsdbProvider(ctrl)
 	container.Opentsdb = opentsdbMock
 
 	var httpMock *service.MockHTTP
